requirements.txt
discord.py
aiohttp
requests
orjson
psutil
colorthief
flask
